Expose SolverEnum and SolverAttribute in API
