+++ CMakeLists.txt
@@ -24,6 +24,7 @@ add_library(smt-switch "${SMT_SWITCH_LIB_TYPE}"
   "${PROJECT_SOURCE_DIR}/src/sort.cpp"
   "${PROJECT_SOURCE_DIR}/src/term.cpp"
   "${PROJECT_SOURCE_DIR}/src/solver.cpp"
+  "${PROJECT_SOURCE_DIR}/src/solver_enums.cpp"
   "${PROJECT_SOURCE_DIR}/src/term_translator.cpp"
   "${PROJECT_SOURCE_DIR}/src/identity_walker.cpp"
   "${PROJECT_SOURCE_DIR}/src/term_hashtable.cpp"
